+++ src/csvpack/__init__.py
@@ -19,7 +19,9 @@ from .common.errors import (
 from .common.metadata import CSVPackMeta
 from .csvlib import (
     CSVReader,
+    CsvReaderIterator,
     CSVWriter,
+    CsvWriterIterator,
     RustCsvReader,
     RustCsvWriter,
 )
@@ -41,9 +43,11 @@ __all__ = (
     "CSVPackValueError",
     "CSVPackWriter",
     "CSVReader",
+    "CsvReaderIterator",
     "CSVWriter",
+    "CsvWriterIterator",
     "RustCsvReader",
     "RustCsvWriter",
 )
 __author__ = "0xMihalich"
-__version__ = "0.1.0.dev2"
+__version__ = "0.1.0.dev3"